Protect QML profiles as well as core
Moves QWebEngineBrowserContext to core and makes use of it from both
widget and qml.
